+++ mirrord/operator/src/setup.rs
@@ -407,13 +407,17 @@ impl OperatorRole {
                         "deployments/scale".to_owned(),
                         "jobs".to_owned(),
                         "rollouts".to_owned(),
+                        "rollouts/scale".to_owned(),
                     ]),
                     verbs: vec!["get".to_owned(), "list".to_owned(), "watch".to_owned()],
                     ..Default::default()
                 },
                 PolicyRule {
-                    api_groups: Some(vec!["apps".to_owned()]),
-                    resources: Some(vec!["deployments/scale".to_owned()]),
+                    api_groups: Some(vec!["apps".to_owned(), "argoproj.io".to_owned()]),
+                    resources: Some(vec![
+                        "deployments/scale".to_owned(),
+                        "rollouts/scale".to_owned(),
+                    ]),
                     verbs: vec!["patch".to_owned()],
                     ..Default::default()
                 },